+++ test/clindex/scanner_test.clj
@@ -49,8 +49,8 @@
                                      (clojure.core/seq
                                       (clojure.core/concat
                                        (clojure.core/list 'clojure.core/+)
-                                       (clojure.core/list 'user/a)
-                                       (clojure.core/list 'user/b))))),
+                                       (clojure.core/list 'test-code/a)
+                                       (clojure.core/list 'test-code/b))))),
                                   :form-str "(defmacro some-macro [a b]\n  `(+ a b)))"}
                                  {:form-list
                                   (defn